# web/src/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ledger": {
      "get": {
        "operationId": "LedgerController_listCells",
        "parameters": [
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Comma-separated fields to include; `_id` is always returned.",
            "schema": {
              "example": "name,status",
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of documents to skip (offset).",
            "schema": {
              "minimum": 0,
              "default": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of documents to return.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "entity",
            "required": false,
            "in": "query",
            "description": "Entity name or a registered alias of one. Resolved to its canonical name first.",
            "schema": {
              "example": "Northgate Business Park",
              "type": "string"
            }
          },
          {
            "name": "measure",
            "required": false,
            "in": "query",
            "description": "Measure slug to filter by.",
            "schema": {
              "example": "cap_rate",
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "description": "Resolved cell state to filter by.",
            "schema": {
              "example": "conflicted",
              "type": "string",
              "enum": [
                "single",
                "adjudicated",
                "conflicted",
                "unknown"
              ]
            }
          },
          {
            "name": "period",
            "required": false,
            "in": "query",
            "description": "Period the cell covers. Absent means the undated period.",
            "schema": {
              "example": "2025-Q1",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Field to sort by. Defaults to entity.",
            "schema": {
              "example": "entity",
              "type": "string",
              "enum": [
                "entity",
                "measure",
                "period"
              ]
            }
          },
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "description": "Sort direction. Defaults to asc.",
            "schema": {
              "example": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The tenant's ledger cells, one per entity × measure × period, each resolved to a state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerCellListResponseDto"
                },
                "examples": {
                  "example": {
                    "summary": "A settled cell and an adjudicated one",
                    "value": {
                      "docs": [
                        {
                          "entity": "Northgate Business Park",
                          "measure": "cap_rate",
                          "period": "2025-Q1",
                          "state": "single",
                          "value": {
                            "amount": 5.25,
                            "unit": "percent",
                            "canonicalAmount": 0.0525
                          },
                          "factIds": [
                            "65f1c2e4a1b2c3d4e5f6a7b8"
                          ]
                        },
                        {
                          "entity": "Cedar Bluff Logistics",
                          "measure": "net_operating_income",
                          "period": "2025",
                          "state": "adjudicated",
                          "value": {
                            "amount": 2134450,
                            "unit": "usd",
                            "canonicalAmount": 2134450
                          },
                          "factIds": [
                            "65f1c2e4a1b2c3d4e5f6a7d1",
                            "65f1c2e4a1b2c3d4e5f6a7d2"
                          ],
                          "conflictId": "65f1c2e4a1b2c3d4e5f6a7e0",
                          "decision": {
                            "conflictId": "65f1c2e4a1b2c3d4e5f6a7e0",
                            "outcome": "resolved",
                            "winningFactId": "65f1c2e4a1b2c3d4e5f6a7d1",
                            "decidedBy": "65f1c2e4a1b2c3d4e5f6a7c9",
                            "reason": "The audited statement supersedes the broker summary.",
                            "resolvedAt": "2026-07-04T09:15:00.000Z",
                            "ruleFired": "authority",
                            "followedProposal": true
                          },
                          "winnerWithdrawn": false
                        }
                      ],
                      "count": 2
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ledger"
        ]
      }
    },
    "/ledger/resolve": {
      "get": {
        "operationId": "LedgerController_resolve",
        "parameters": [
          {
            "name": "entity",
            "required": true,
            "in": "query",
            "description": "Entity name or a registered alias of one. Resolved to its canonical name first.",
            "schema": {
              "example": "Northgate Business Park",
              "type": "string"
            }
          },
          {
            "name": "measure",
            "required": true,
            "in": "query",
            "description": "Measure slug.",
            "schema": {
              "example": "cap_rate",
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": false,
            "in": "query",
            "description": "Period the cell covers. Absent resolves the undated period.",
            "schema": {
              "example": "2025-Q1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One cell resolved in full, with the citations the resolution rests on.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerResolutionResponseDto"
                },
                "examples": {
                  "example": {
                    "summary": "A single agreed value with its citation",
                    "value": {
                      "entity": "Northgate Business Park",
                      "measure": "cap_rate",
                      "period": "2025-Q1",
                      "state": "single",
                      "value": {
                        "amount": 5.25,
                        "unit": "percent",
                        "canonicalAmount": 0.0525
                      },
                      "factIds": [
                        "65f1c2e4a1b2c3d4e5f6a7b8"
                      ],
                      "citations": [
                        {
                          "factId": "65f1c2e4a1b2c3d4e5f6a7b8",
                          "documentId": "65f1c2e4a1b2c3d4e5f6a7c0",
                          "documentVersionId": "65f1c2e4a1b2c3d4e5f6a7c1",
                          "sha256": "9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
                          "locator": {
                            "kind": "xlsx-cell",
                            "sheet": "Comps",
                            "cell": "D4"
                          },
                          "extractorVersion": "xlsx-1",
                          "quote": "5.25%",
                          "withdrawn": false
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No measure with that slug exists for the tenant, or it has been rejected.",
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Unknown measure",
                    "value": {
                      "statusCode": 404,
                      "message": "Measure 'occupancy' not found",
                      "error": "Not Found"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ledger"
        ]
      }
    },
    "/ledger/facts": {
      "get": {
        "operationId": "LedgerController_listFacts",
        "parameters": [
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Comma-separated fields to include; `_id` is always returned.",
            "schema": {
              "example": "name,status",
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of documents to skip (offset).",
            "schema": {
              "minimum": 0,
              "default": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of documents to return.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "entity",
            "required": true,
            "in": "query",
            "description": "Entity name or a registered alias of one. Resolved to its canonical name first.",
            "schema": {
              "example": "Northgate Business Park",
              "type": "string"
            }
          },
          {
            "name": "measure",
            "required": true,
            "in": "query",
            "description": "Measure slug.",
            "schema": {
              "example": "cap_rate",
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": false,
            "in": "query",
            "description": "Period the cell covers. Absent means the undated period.",
            "schema": {
              "example": "2025-Q1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Every fact behind one cell, at any measure status — the drill-down deliberately shows proposed-measure facts that the cell view excludes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactListResponseDto"
                },
                "examples": {
                  "example": {
                    "summary": "One confirmed fact behind a cell",
                    "value": {
                      "docs": [
                        {
                          "id": "65f1c2e4a1b2c3d4e5f6a7b8",
                          "factKey": {
                            "entity": "Northgate Business Park",
                            "metric": "cap_rate",
                            "period": "2025-Q1"
                          },
                          "value": {
                            "amount": 5.25,
                            "unit": "percent"
                          },
                          "canonicalAmount": 0.0525,
                          "rawText": "5.25%",
                          "confidence": 0.95,
                          "extractionMethod": "regex",
                          "measureId": "65f1c2e4a1b2c3d4e5f6a7a1",
                          "measureVersion": 1,
                          "measureStatus": "confirmed",
                          "periodStart": "2025-01-01T00:00:00.000Z",
                          "periodEnd": "2025-03-31T00:00:00.000Z",
                          "entityMatched": true,
                          "citation": {
                            "factId": "65f1c2e4a1b2c3d4e5f6a7b8",
                            "documentId": "65f1c2e4a1b2c3d4e5f6a7c0",
                            "documentVersionId": "65f1c2e4a1b2c3d4e5f6a7c1",
                            "sha256": "9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
                            "locator": {
                              "kind": "xlsx-cell",
                              "sheet": "Comps",
                              "cell": "D4"
                            },
                            "extractorVersion": "xlsx-1",
                            "quote": "5.25%",
                            "withdrawn": false
                          },
                          "withdrawn": false,
                          "superseded": false,
                          "createdAt": "2026-07-01T00:00:00.000Z"
                        }
                      ],
                      "count": 1
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No measure with that slug exists for the tenant, or it has been rejected.",
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Unknown measure",
                    "value": {
                      "statusCode": 404,
                      "message": "Measure 'occupancy' not found",
                      "error": "Not Found"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ledger"
        ]
      }
    },
    "/ledger/entities": {
      "get": {
        "operationId": "LedgerController_listEntities",
        "parameters": [
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Comma-separated fields to include; `_id` is always returned.",
            "schema": {
              "example": "name,status",
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of documents to skip (offset).",
            "schema": {
              "minimum": 0,
              "default": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of documents to return.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entities the ledger holds confirmed facts for.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerEntityListResponseDto"
                },
                "examples": {
                  "example": {
                    "summary": "Two entities",
                    "value": {
                      "docs": [
                        {
                          "entity": "Cedar Bluff Logistics",
                          "factCount": 8,
                          "measureCount": 3
                        },
                        {
                          "entity": "Northgate Business Park",
                          "factCount": 12,
                          "measureCount": 4
                        }
                      ],
                      "count": 2
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ledger"
        ]
      }
    },
...
  },
  "info": {
    "title": "Evidence Ops API",
    "description": "Evidence Ops API specification",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "LedgerCellResponseDto": {
        "type": "object",
        "properties": {
          "entity": {
            "type": "string",
            "example": "Northgate Business Park",
            "description": "Canonical entity name."
          },
          "measure": {
            "type": "string",
            "example": "cap_rate",
            "description": "Measure slug."
          },
          "period": {
            "type": "string",
            "example": "2025-Q1",
            "description": "Period key the cell covers; 'undated' when the facts carry no period."
          },
          "state": {
            "type": "string",
            "example": "single",
            "enum": [
              "single",
              "adjudicated",
              "conflicted",
              "unknown"
            ],
            "description": "How the record resolves this cell: one agreed value, a human-adjudicated winner, an open disagreement, or nothing known."
          },
          "value": {
            "type": "object",
            "description": "The resolved value. Absent for a conflicted or unknown cell."
          },
          "factIds": {
            "description": "Facts behind this cell — the disagreeing set when the cell is conflicted.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "conflictId": {
            "type": "string",
            "description": "The conflict over this cell, when one has been recorded."
          },
          "decision": {
            "type": "object",
            "description": "The decision record behind an adjudicated cell."
          },
          "winnerWithdrawn": {
            "type": "boolean",
            "description": "True when an adjudicated winner sits on a withdrawn document version. The decision still stands; the document behind it no longer does."
          }
        },
        "required": [
          "entity",
          "measure",
          "period",
          "state",
          "factIds"
        ]
      },
      "LedgerCellListResponseDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "example": 2,
            "description": "Total number of documents returned."
          },
          "docs": {
            "description": "Resolved ledger cells.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LedgerCellResponseDto"
            }
          }
        },
        "required": [
          "count",
          "docs"
        ]
      },
      "LedgerResolutionResponseDto": {
        "type": "object",
        "properties": {
          "entity": {
            "type": "string",
            "example": "Northgate Business Park",
            "description": "Canonical entity name."
          },
          "measure": {
            "type": "string",
            "example": "cap_rate",
            "description": "Measure slug."
          },
          "period": {
            "type": "string",
            "example": "2025-Q1",
            "description": "Period key the cell covers; 'undated' when the facts carry no period."
          },
          "state": {
            "type": "string",
            "example": "single",
            "enum": [
              "single",
              "adjudicated",
              "conflicted",
              "unknown"
            ],
            "description": "How the record resolves this cell: one agreed value, a human-adjudicated winner, an open disagreement, or nothing known."
          },
          "value": {
            "type": "object",
            "description": "The resolved value. Absent for a conflicted or unknown cell."
          },
          "factIds": {
            "description": "Facts behind this cell — the disagreeing set when the cell is conflicted.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "conflictId": {
            "type": "string",
            "description": "The conflict over this cell, when one has been recorded."
          },
          "decision": {
            "type": "object",
            "description": "The decision record behind an adjudicated cell."
          },
          "winnerWithdrawn": {
            "type": "boolean",
            "description": "True when an adjudicated winner sits on a withdrawn document version. The decision still stands; the document behind it no longer does."
          },
          "citations": {
            "description": "Citations for the facts this resolution rests on, one per fact.",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "entity",
          "measure",
          "period",
          "state",
          "factIds",
          "citations"
        ]
      },
      "FactResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "65f1c2e4a1b2c3d4e5f6a7b8",
            "description": "Fact identifier."
          },
          "factKey": {
            "type": "object",
            "description": "Entity, metric and period this fact is keyed by."
          },
          "value": {
            "type": "object",
            "description": "The extracted amount and its unit."
          },
          "canonicalAmount": {
            "type": "number",
            "description": "The value in the measure's canonical unit. Absent when the fact's unit does not convert."
          },
          "rawText": {
            "type": "string",
            "description": "The source text this fact was read from."
          },
          "confidence": {
            "type": "number",
            "example": 0.92,
            "description": "Extractor confidence in this fact."
          },
          "extractionMethod": {
            "type": "string",
            "example": "regex",
            "enum": [
              "llm",
              "regex",
              "manual"
            ],
            "description": "How this fact was extracted."
          },
          "measureId": {
            "type": "string",
            "description": "The measure row this fact was stamped under."
          },
          "measureVersion": {
            "type": "number",
            "example": 1,
            "description": "The measure version in force at extraction time — provenance, never re-derived from the measure as it stands now."
          },
          "measureStatus": {
            "type": "string",
            "example": "confirmed",
            "enum": [
              "proposed",
              "confirmed"
            ],
            "description": "Whether the measure behind this fact is confirmed. A proposed-measure fact is stored but excluded from conflict detection and ledger answers."
          },
          "periodStart": {
            "format": "date-time",
            "type": "string",
            "description": "Start of the period this fact covers."
          },
          "periodEnd": {
            "format": "date-time",
            "type": "string",
            "description": "End of the period this fact covers."
          },
          "observedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the fact was observed, where stated."
          },
          "entityMatched": {
            "type": "boolean",
            "description": "Whether the fact's entity matched a registered canonical entity."
          },
          "citation": {
            "type": "object",
            "description": "Where this fact can be re-checked. Absent when the document version behind it could not be resolved."
          },
          "withdrawn": {
            "type": "boolean",
            "description": "Whether the document version behind this fact has been withdrawn."
          },
          "superseded": {
            "type": "boolean",
            "description": "Whether a newer version of the same document has superseded this one."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the fact was extracted."
          }
        },
        "required": [
          "id",
          "factKey",
          "value",
          "rawText",
          "confidence",
          "extractionMethod",
          "measureId",
          "measureVersion",
          "measureStatus",
          "createdAt"
        ]
      },
      "FactListResponseDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "example": 2,
            "description": "Total number of documents returned."
          },
          "docs": {
            "description": "Facts behind the addressed cell.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FactResponseDto"
            }
          }
        },
        "required": [
          "count",
          "docs"
        ]
      },
      "LedgerEntityResponseDto": {
        "type": "object",
        "properties": {
          "entity": {
            "type": "string",
            "example": "Northgate Business Park",
            "description": "Canonical entity name."
          },
          "factCount": {
            "type": "number",
            "example": 12,
            "description": "Confirmed-measure facts recorded against this entity."
          },
          "measureCount": {
            "type": "number",
            "example": 4,
            "description": "Distinct measures this entity has facts for."
          }
        },
        "required": [
          "entity",
          "factCount",
          "measureCount"
        ]
      },
      "LedgerEntityListResponseDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "example": 2,
            "description": "Total number of documents returned."
          },
          "docs": {
            "description": "Entities the ledger holds confirmed facts for.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LedgerEntityResponseDto"
            }
          }
        },
        "required": [
          "count",
          "docs"
        ]
      },
...
    }
  }
}